refactor: split cli helpers into srp modules
